move cli to its own crate
