+++ Cargo.toml
@@ -8,23 +8,15 @@ authors = ["Brock Smedley"]
 name = "contender_core"
 path = "src/lib.rs"
 
-[[bin]]
-name = "contender_cli"
-path = "src/bin/cli.rs"
-
 [dependencies]
 alloy = { version = "0.3.6", features = ["full", "node-bindings"] }
-clap = { version = "4.5.16", features = ["derive"] }
 eyre = "0.6.12"
-lazy_static = "1.5.0"
 rand = "0.8.5"
 serde = "1.0.209"
-serde_json = "1.0.128"
-tokio = { version = "1.40.0", features = ["rt-multi-thread"] }
+tokio = { version = "1.40.0" }
 toml = "0.8.19"
 r2d2_sqlite = "0.25.0"
 rusqlite = "0.32.1"
 r2d2 = "0.8.10"
 futures = "0.3.30"
 async-trait = "0.1.82"
-csv = "1.3.0"
