+++ cli/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "contender_cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap = { version = "4.5.16", features = ["derive"] }
+tokio = { version = "1.40.0", features = ["rt-multi-thread"] }
+alloy = { version = "0.3.6", features = ["full", "node-bindings"] }
+csv = "1.3.0"
+contender = { path = "../" }
