+++ cli/src/commands.rs
